coordinates_var_precision is now a param of pset
By default it is single precision for A-grid U field and double precision for
C-grid

+++ tests/customed_header.h
@@ -1,5 +1,5 @@
 
-static inline void func(CField *f, double *lon, double *dt)
+static inline void func(CField *f, double *lon, float *dt)
 {
   float (*data)[f->xdim] = (float (*)[f->xdim]) f->data;
   float u = data[2][1];
